Remove data.py, remove max_restarts
