+++ examples/computer_vision/cifar10_tf_keras/const.yaml
@@ -17,5 +17,4 @@ searcher:
     epochs: 32
 min_validation_period:
   epochs: 1
-max_restarts: 0
 entrypoint: model_def:CIFARTrial
